+++ src/main/java/com/supermartijn642/connectedglass/CGGlassBlock.java
@@ -3,7 +3,7 @@ package com.supermartijn642.connectedglass;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.level.block.AbstractGlassBlock;
 import net.minecraft.world.level.block.SoundType;
-import net.minecraft.world.level.material.Material;
+import net.minecraft.world.level.block.state.properties.NoteBlockInstrument;
 
 /**
  * Created 5/7/2020 by SuperMartijn642
@@ -14,7 +14,7 @@ public class CGGlassBlock extends AbstractGlassBlock {
     public final boolean connected;
 
     public CGGlassBlock(String texture, boolean connected){
-        super(Properties.of(Material.GLASS).sound(SoundType.GLASS).strength(0.3f).noOcclusion());
+        super(Properties.of().sound(SoundType.GLASS).instrument(NoteBlockInstrument.HAT).strength(0.3f).noOcclusion().isValidSpawn((a, b, c, d) -> false).isRedstoneConductor((a, b, c) -> false).isSuffocating((a, b, c) -> false).isViewBlocking((a, b, c) -> false));
         this.texture = new ResourceLocation("connectedglass", texture);
         this.connected = connected;
     }