+++ changelog.md
@@ -1,2 +1,2 @@
 ### Connected Glass 1.1.6
-- Initial release of Connected Glass for Fabric
+- Initial release of Connected Glass for Minecraft 1.20